feat: get PRIVATE_KEY from SST.Config
Uses SST.Config to get our base64 encoded EdSigner keypair for our service.

`Config` magically loads properties with fallbacks from aws SSM. I've refactored things so that we don't use it in tests to keep things simpler.

Note: We can use Config in tests, but it means our tests would call out to SSM, and fetch a fallback value for any secrets we use, so i've opted for the conceptually cleaner option of only using `Config` magic at runtime, and not in tests.

License: MIT
