+++ api/test/helpers/ucanto.js
@@ -2,7 +2,7 @@ import * as UcantoClient from '@ucanto/client'
 import { CAR, CBOR } from '@ucanto/transport'
 import * as Signer from '@ucanto/principal/ed25519'
 
-import { createUcantoServer } from '../../functions/ucan-invocation-router.js'
+import { createUcantoServer } from '../../service/index.js'
 import { createCarStore } from '../../buckets/car-store.js'
 import { createStoreTable } from '../../tables/store.js'
 import { createUploadTable } from '../../tables/upload.js'
@@ -10,24 +10,28 @@ import { createSigner } from '../../signer.js'
 
 import { getSigningOptions } from '../utils.js'
 
+/** @typedef {import('@ucanto/interface').Principal} Principal */
+
 /**
+ * @param {Principal} service 
  * @param {any} ctx
  */
-export function createTestingUcantoServer(ctx) {
- return createUcantoServer({
-   storeTable: createStoreTable(ctx.region, ctx.tableName, {
-     endpoint: ctx.dbEndpoint
-   }),
-   uploadTable: createUploadTable(ctx.region, ctx.tableName, {
-     endpoint: ctx.dbEndpoint
-   }),
-   carStoreBucket: createCarStore(ctx.region, ctx.bucketName, { ...ctx.s3ClientOpts }),
-   signer: createSigner(getSigningOptions(ctx))
+export async function createTestingUcantoServer(service, ctx) {
+  return createUcantoServer({
+    serviceSigner: service,
+    storeTable: createStoreTable(ctx.region, ctx.tableName, {
+      endpoint: ctx.dbEndpoint
+    }),
+    uploadTable: createUploadTable(ctx.region, ctx.tableName, {
+      endpoint: ctx.dbEndpoint
+    }),
+    carStoreBucket: createCarStore(ctx.region, ctx.bucketName, { ...ctx.s3ClientOpts }),
+    signer: createSigner(getSigningOptions(ctx))
  })
 }
 
 /**
- * @param {import('@ucanto/principal/ed25519').EdSigner} service 
+ * @param {Principal} service 
  * @param {any} context 
  * @returns 
  */
@@ -36,12 +40,12 @@ export async function getClientConnection (service, context) {
     id: service,
     encoder: CAR,
     decoder: CBOR,
-    channel: await createTestingUcantoServer(context),
+    channel: await createTestingUcantoServer(service, context),
   })
 }
 
 /**
- * @param {import("@ucanto/principal/dist/src/ed25519/type.js").EdSigner<"key">} audience
+ * @param {Principal} audience
  */
 export async function createSpace (audience) {
   const space = await Signer.generate()
